+++ CHANGELOG.md
@@ -24,6 +24,10 @@ and this project follows versions of format `{year}.{month}.{patch_number}`.
   - `/bookings/{booking_id}/cancel` (POST)
 - BREAKING: Added authentication of requests from MSS by verifying that `x-mss-signature` is signed using MSS's private key
 - Added canceling of jobs via MSS
+- Added and adapted SpiDAC class from autocalibration package for coupler biasing
+- Modify Quantify quantum executor by calling SpiDAC to bring up currents before running a schedule
+- Add a placeholder Wacqt-two-qubit-gate function for Qobj to QuantifySchedule mapper with a flux port reference 
+- Add safety features to SpiDAC to limit current values and bring-up/down speed
 - Add filtering bookings by minimum/maximum start timestamps
 
 ### Changed